requirements.txt
brotlipy==0.7.0
Flask-Bootstrap==3.3.7.1
Flask-WTF==0.14.3
matplotlib==3.2.1
pipdeptree==0.13.2
pyOpenSSL==19.1.0
PySocks==1.7.1
pytube3==9.6.4
rq==1.4.2
spleeter==1.5.2
youtube-search==0.1.3
